backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase==2.4.1
upstash-redis==1.0.0
apscheduler==3.10.4
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
httpx==0.26.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
shapely==2.0.3
python-multipart==0.0.9
